pyproject.toml
[project]
name = "practical-25-winter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.45",
    "statsmodels>=0.14.6",
]
